Rewrited from jpaApi inject to entityManager factory

+++ app/dao/SecurityRoleDao.java
@@ -9,6 +9,7 @@ import java.io.Serializable;
 
 /**
  * Created by mik on 11.09.2016.
+ *
  */
 public class SecurityRoleDao extends GenericDaoImpl<SecurityRole, Serializable> {
     public SecurityRoleDao(EntityManager em) {
@@ -19,7 +20,7 @@ public class SecurityRoleDao extends GenericDaoImpl<SecurityRole, Serializable> 
         return (SecurityRole)getSession().createCriteria(SecurityRole.class).add(Restrictions.eq("role_name", roleName)).list().get(0);
     }
 
-    public int count() {
-        return (int)getSession().createCriteria(SecurityRole.class).setProjection(Projections.rowCount()).uniqueResult();
+    public Long count() {
+        return (Long)getSession().createCriteria(SecurityRole.class).setProjection(Projections.rowCount()).uniqueResult();
     }
 }
